+++ src/components/shared/index.ts
@@ -1,0 +1,3 @@
+export { PublicHeader } from './PublicHeader'
+export { PublicFooter } from './PublicFooter'
+export { PWAInstallBanner } from './PWAInstallBanner'